requirements.txt
google-genai
flask
flask-cors
python-dotenv

